pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "a11y-autofix"
version = "2.0.0"
description = "Sistema de detecção e correção automática de acessibilidade com experimentação multi-modelo"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "a11y-autofix contributors" }]
keywords = ["accessibility", "wcag", "a11y", "react", "typescript", "llm"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "structlog>=24.1.0",
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "rich-click>=1.7.0",
    "httpx[http2]>=0.27.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",
    "tenacity>=8.3.0",
    "openai>=1.30.0",
    "playwright>=1.44.0",
    "orjson>=3.10.0",
    "deepdiff>=7.0.0",
    "tabulate>=0.9.0",
    "gitpython>=3.1.43",
    "python-dotenv>=1.0.1",
    "jinja2>=3.1.4",
    "pyyaml>=6.0.1",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-benchmark>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-pyyaml",
    "types-tabulate",
    "types-python-dateutil",
]

[project.scripts]
a11y-autofix = "a11y_autofix.cli:app"

[project.urls]
Repository = "https://github.com/seu-usuario/a11y-autofix"
Documentation = "https://github.com/seu-usuario/a11y-autofix/blob/main/docs/"

[tool.hatch.build.targets.wheel]
packages = ["a11y_autofix"]

# ─── Ruff ───────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "N", "ANN", "S", "A", "COM", "PT"]
ignore = [
    "ANN101", "ANN102", "ANN401",
    "S101",   # allow assert in tests
    "S603",   # subprocess (needed for CLI tools)
    "S607",   # partial path in subprocess
    "COM812", # trailing comma (conflicts with formatter)
    "B008",   # do not perform function calls in default arguments
]
per-file-ignores = { "tests/**/*.py" = ["ANN", "S"] }

[tool.ruff.format]
quote-style = "double"

# ─── MyPy ────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
exclude = ["tests/", "scripts/"]

# ─── Pytest ──────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--tb=short -v"
markers = [
    "integration: integration tests (require external services)",
    "slow: slow tests",
]

[tool.coverage.run]
source = ["a11y_autofix"]
omit = ["tests/*", "scripts/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
